requirements.txt
# Enhanced Food Detection System Requirements
tensorflow>=2.10.0
opencv-python>=4.6.0
pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
beautifulsoup4>=4.11.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
